Updated lists with styles
