+++ src/components/albums/index.ts
@@ -1,0 +1,3 @@
+export * from './albumListItem';
+export * from './albumsList';
+
