+++ src/components/common/index.ts
@@ -1,3 +1,3 @@
 export * from './header';
 export * from './footer';
-export * from './albumListItem';
+export * from '../albums/albumListItem';
